[Fix] Improve validation for "fill in the blank"
ISSUE:
On latest version available of quiz_master_next, if you try to validate text on other non-english language you face a validation error forcing correct answers to be added more than once. For example, if correct answer is "Atún", is necessary to add another correct answer all capitalized "ATÚN" in case user who fills the quiz use capital letters. This because to validate we transform text to uppercase letters, but this works well if you don't use special characters like "ü" or "á", and so on...

Also, when user who fills the quiz use autocorrect an extra whitespace is added and if is not deleted then answer is validated as FALSE.

SOLUTION:
1.- For uppercase letters validation replace "strtoupper" for "mb_strtoupper", that way all letters (no matter if special char or not) will be converted to uppercase.
2.- For whitespace, we need to decode the correct text and user text and then replace all white spaces: the ones added at the beginning and the end will be removed, and if more than one whitespace is added between 2 words will be replaced for only one whitespace.