Hi Developer fixed the user comment not showing in results page.
Earlier when the user has commented on a test the Admin was unable to view the comment cause it was hidden because the comment option is optional(Just an assumption). So we modified the code and add a conditional statement which checks if the user has wrote a comment.

$results variable used, checks the unserialized data and checks if the comment index exist which in our case is 2. If exist displays the user comment else hides the comment div.

Thanks