Now the comment area will be shown if comments are enabled.
We have modfified the admin-results-details-page.php. I have wrote a sql query to get the data regarding the quiz the results page belongs too. You can see that through git show {current_commit_id}.

We unserialize quiz_data->quiz_setting and store it in a variable $unserialized_quiz_setting.

All the data in quiz_Setting is serialized. Then we again unserialized the
$unserialized_quiz_setting['quiz_options'] and store it in $comments_enabled variable.

The values for comments enabled or not is set like this. '0' for enable and '1' for disabled(quite opposite what I thought). We check $comments_enabled['comment_section'] === '0' and display the div accordingly.