Recaptcha is now available for randomized quizzes
Earlier recaptcha was hidden when questions are randomized. Cause earlier to show recaptcha newer pagination settings were used.
We checked the difference between the paginations between randomized and non-ranzomized quizzes and found that ranzomized quizzed don't have .qsm-page class.

So by refrencing the existence of that class we trigger recaptcha for randomized quizzes