A Question Per category feature is added to QSM plugin. This feature allows the user to choose a specific amount of questions per category. This is how it works:-
Gets all the categories the user has selected.
Runs a loop For each individual category and gets question_per_category amount of questions.
Merge all the questions to a array named $big_array.
Checks if count($big_array) is more than the question_from_total, If it's then it runs a loop and get either random entries or serial entries depending on the setting and adds them to $question array.
else if it's less we run a similar loop configured for less questions. And we return the array

Eg:-
If question_from_total = 5 and question_per_category = 2 and the selected categories is 3 and for all the categories user have created 2 questions.
Then: get 2 questions from each category: 2*3 = 6 questions
6 is bigger than 5 so we get random(depends on setting) 5 questions from those 6 questions
if we get a number less than 5 we just return it or randomize it.